refactor: 💡 Seperate quotable design const and interface data
