+++ libs/interfaces/src/lib/design/quotable/index.ts
@@ -1,3 +1,4 @@
-export * from './quotable-design-info';
+export * from './quotable-design-info.const';
+export * from './quotable-design-info.interface';
 export * from './quotable.design-type.enum';
 export * from './quotable.interface';
