+++ libs/interfaces/src/lib/design/quotable/quotable-design-info.const.ts
@@ -1,12 +1,6 @@
 import { QuotableDesignEnum, QuotableDesignNameDanishEnum, QuotableDesignNameEnglishEnum } from '@assets';
+import { QuotableFrameInfo } from './quotable-design-info.interface';
 import { QuotableType } from './quotable.design-type.enum';
-
-export interface QuotableFrameInfo {
-  src: string;
-  nameDk: string;
-  nameEn: string;
-  productType: QuotableType[];
-}
 
 export const quotableFrames: QuotableFrameInfo[] = [
   {
