+++ libs/interfaces/src/lib/design/quotable/quotable-design-info.interface.ts
@@ -1,0 +1,8 @@
+import { QuotableType } from './quotable.design-type.enum';
+
+export interface QuotableFrameInfo {
+  src: string;
+  nameDk: string;
+  nameEn: string;
+  productType: QuotableType[];
+}
